+++ CHANGELOG.md
@@ -26,6 +26,8 @@ See for more information about XPrivacy 3 [this FAQ](https://github.com/M66B/XPr
 **Next release**
 
 * Fixed all problems reported through the debug info
+* Updated Android support library
+* Using Android SDK CardView library
 * Updated Norwegian translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)