+++ CHANGELOG.md
@@ -27,8 +27,9 @@ See for more information about XPrivacy 3 [this FAQ](https://github.com/M66B/XPr
 **Next release**
 
 * Fixed all problems reported through the debug info
-* Updated Android support library
+* Updated Android support library project
 * Using Android SDK CardView library
+* Material design
 * Updated Norwegian translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)