+++ CHANGELOG.md
@@ -26,14 +26,16 @@ See for more information about XPrivacy 3 [this FAQ](https://github.com/M66B/XPr
 
 **Next release**
 
+[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
+
+**Version 3.4.2 TEST**
+
 * Fixed all problems reported through the debug info
 * Updated Android support library project
 * Using Android SDK CardView library
 * Material design
 * Use subtitle for operation name
 * Updated Norwegian translation
-
-[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 3.4.1 BETA**